rename: userInfo api 네이밍 변경
